small bugfix
Former-commit-id: 71a43e7670153021d6ea1fcb778ed43364f38bd3

+++ py4DSTEM/io/datastructure/datacube.py
@@ -97,7 +97,7 @@ class DataCube(DataObject):
         self = preprocess.bin_data_diffraction(self, bin_factor)
 
     def bin_data_mmap(self, bin_factor, dtype=np.float32):
-        self = preprocess.bin_data_mmap(self, bin_factor)
+        self = preprocess.bin_data_mmap(self, bin_factor, dtype=dtype)
 
     def bin_data_real(self, bin_factor):
         self = preprocess.bin_data_real(self, bin_factor)
